+++ src/t8_schemes/t8_default/t8_default_pyramid/t8_dpyramid_bits.c
@@ -1745,7 +1745,7 @@ t8_dpyramid_vertex_reference_coords (const t8_dpyramid_t *elem,
 void
 t8_dpyramid_compute_reference_coords (const t8_dpyramid_t *elem,
                                       const double *ref_coords,
-                                      const int num_coords,
+                                      const size_t num_coords,
                                       double *out_coords)
 {
   T8_ASSERT (ref_coords != NULL);
